+++ mistralrs-core/src/vision_models/phi3_inputs_processor.rs
